src/pyproject.toml
[tool.poetry]
name = "test_evraz"
version = "0.1.0"
description = ""
authors = ["Ra1ze505 <87398772+Ra1ze505@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
openpyxl = "^3.0.10"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
